requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers